+++ crate/clients/ckms/src/tests/kms/attributes/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod test_modify_attribute;
+
+pub(crate) const SUB_COMMAND: &str = "attributes";